TimeSeriesBase: new common base class for TimeSeries and StateVector
    - TimeSeriesBase defines all common functionality
    - gwpy.timeseries.timeseries.TimeSeries now defines all Fourier
      methods, for example
    - means StateVector and StateTimeSeries aren't bloatet with methods
      they can't use
