+++ gwpy/timeseries/__init__.py
@@ -25,5 +25,6 @@ __author__ = "Duncan Macleod <duncan.macleod@ligo.org>"
 __version__ = version.version
 
 from .core import *
+from .timeseries import *
 from .statevector import *
 from .io import *
